Split up ConsentBannerAndConsentManagment into multiple files
