+++ src/gdpr/ConsentBannerAndConsentManagement/ConsentConsentManagement.tsx
@@ -1,192 +1,21 @@
-import React, { useReducer, useEffect, useMemo, useState, useId, type ReactNode } from "react";
-import { fr } from "../fr";
-import { createComponentI18nApi } from "../i18n";
-import { getLink, type RegisteredLinkProps } from "../link";
-import { createModal } from "../Modal";
-import type { ExtractFinalityFromFinalityDescription, FinalityConsent } from "./types";
-import type { ProcessConsentChanges } from "./processConsentChanges";
-import { FooterBottomItem } from "../Footer";
-import { useLang } from "../i18n";
-import { useIsModalOpen } from "../Modal/useIsModalOpen";
-import { isBrowser } from "../tools/isBrowser";
+import React, { useEffect, useMemo, useState, useId, type ReactNode } from "react";
+import { fr } from "../../fr";
+import type { RegisteredLinkProps } from "../../link";
+import { createModal } from "../../Modal";
+import type { ExtractFinalityFromFinalityDescription, FinalityConsent } from "../types";
+import type { ProcessConsentChanges } from "../processConsentChanges";
+import { useLang } from "../../i18n";
+import { useIsModalOpen } from "../../Modal/useIsModalOpen";
 import { objectKeys } from "tsafe/objectKeys";
 import { assert } from "tsafe/assert";
 import {
     createProcessConsentChanges,
     createFullDenyFinalityConsent
-} from "./processConsentChanges";
+} from "../processConsentChanges";
 import { exclude } from "tsafe/exclude";
+import { useTranslation } from "./translation";
 
-export function createConsentBannerAndConsentManagement<
-    FinalityDescription extends Record<
-        string,
-        { title: ReactNode; description?: ReactNode; subFinalities?: Record<string, ReactNode> }
-    >
->(params: {
-    finalityDescription: ((params: { lang: string }) => FinalityDescription) | FinalityDescription;
-    useFinalityConsent: () =>
-        | FinalityConsent<ExtractFinalityFromFinalityDescription<FinalityDescription>>
-        | undefined;
-    processConsentChanges: ProcessConsentChanges<
-        ExtractFinalityFromFinalityDescription<FinalityDescription>
-    >;
-    personalDataPolicyLinkProps?: RegisteredLinkProps;
-    finalities: ExtractFinalityFromFinalityDescription<FinalityDescription>[];
-}) {
-    const {
-        finalityDescription,
-        useFinalityConsent,
-        processConsentChanges,
-        personalDataPolicyLinkProps,
-        finalities
-    } = params;
-
-    const { ConsentManagement, consentModalButtonProps, useIsConsentManagementOpen } =
-        createConsentManagement({
-            finalityDescription,
-            personalDataPolicyLinkProps,
-            useFinalityConsent,
-            processConsentChanges,
-            finalities
-        });
-
-    const { ConsentBanner } = createConsentBanner({
-        personalDataPolicyLinkProps,
-        processConsentChanges,
-        consentModalButtonProps
-    });
-
-    const { FooterConsentManagementItem } = createFooterConsentManagementItem({
-        consentModalButtonProps
-    });
-
-    function ConsentBannerAndConsentManagement() {
-        const [isHydrated, setIsHydrated] = useReducer(() => true, false);
-
-        useEffect(() => {
-            processConsentChanges({ "type": "no changes but trigger consent callbacks" });
-
-            setIsHydrated();
-        }, []);
-
-        const finalityConsent = useFinalityConsent();
-
-        const isConsentManagementOpen = useIsConsentManagementOpen();
-
-        if (!isHydrated) {
-            return null;
-        }
-
-        return (
-            <>
-                {finalityConsent === undefined && !isConsentManagementOpen && <ConsentBanner />}
-                <ConsentManagement />
-            </>
-        );
-    }
-
-    const { FooterPersonalDataPolicyItem } = createFooterPersonalDataPolicyItem({
-        personalDataPolicyLinkProps
-    });
-
-    return {
-        ConsentBannerAndConsentManagement,
-        FooterConsentManagementItem,
-        FooterPersonalDataPolicyItem
-    };
-}
-
-function createConsentBanner<Finality extends string>(params: {
-    personalDataPolicyLinkProps: RegisteredLinkProps | undefined;
-    processConsentChanges: ProcessConsentChanges<Finality>;
-    consentModalButtonProps: {
-        "aria-controls": string;
-        "data-fr-opened": boolean;
-    };
-}) {
-    const { personalDataPolicyLinkProps, processConsentChanges, consentModalButtonProps } = params;
-
-    function ConsentBanner() {
-        const { t } = useTranslation();
-
-        const [hostname, setHostname] = useState("");
-
-        const [isProcessingChanges, setIsProcessingChanges] = useState(false);
-
-        useEffect(() => {
-            if (!isBrowser) {
-                return;
-            }
-
-            setHostname(location.host);
-        }, []);
-
-        return (
-            <>
-                <div className={fr.cx("fr-consent-banner")}>
-                    <h2 className={fr.cx("fr-h6")}>{t("about cookies", { hostname })}</h2>
-                    <div /*className={fr.cx("fr-consent-banner__content")}*/>
-                        <p className={fr.cx("fr-text--sm")}>
-                            {t("welcome message", { personalDataPolicyLinkProps })}
-                        </p>
-                    </div>
-                    <ul
-                        className={fr.cx(
-                            "fr-consent-banner__buttons",
-                            "fr-btns-group",
-                            "fr-btns-group--right",
-                            "fr-btns-group--inline-reverse",
-                            "fr-btns-group--inline-sm"
-                        )}
-                    >
-                        <li>
-                            <button
-                                className={fr.cx("fr-btn")}
-                                title={t("accept all - title")}
-                                onClick={async () => {
-                                    setIsProcessingChanges(true);
-                                    await processConsentChanges({ "type": "grantAll" });
-                                    setIsProcessingChanges(false);
-                                }}
-                                disabled={isProcessingChanges}
-                            >
-                                {t("accept all")}
-                            </button>
-                        </li>
-                        <li>
-                            <button
-                                className={fr.cx("fr-btn")}
-                                title={t("refuse all - title")}
-                                onClick={() => {
-                                    setIsProcessingChanges(true);
-                                    processConsentChanges({ "type": "denyAll" });
-                                    setIsProcessingChanges(false);
-                                }}
-                                disabled={isProcessingChanges}
-                            >
-                                {t("refuse all")}
-                            </button>
-                        </li>
-                        <li>
-                            <button
-                                className={fr.cx("fr-btn", "fr-btn--secondary")}
-                                title={t("customize cookies - title")}
-                                disabled={isProcessingChanges}
-                                {...consentModalButtonProps}
-                            >
-                                {t("customize")}
-                            </button>
-                        </li>
-                    </ul>
-                </div>
-            </>
-        );
-    }
-
-    return { ConsentBanner };
-}
-
-function createConsentManagement<
+export function createConsentManagement<
     FinalityDescription extends Record<
         string,
         { title: ReactNode; description?: ReactNode; subFinalities?: Record<string, ReactNode> }
@@ -601,183 +430,3 @@ function createConsentManagement<
 
     return { ConsentManagement, consentModalButtonProps, useIsConsentManagementOpen };
 }
-
-function createFooterConsentManagementItem(params: {
-    consentModalButtonProps: {
-        "aria-controls": string;
-        "data-fr-opened": boolean;
-    };
-}) {
-    const { consentModalButtonProps } = params;
-
-    function FooterConsentManagementItem() {
-        const { t } = useTranslation();
-
-        return (
-            <FooterBottomItem
-                bottomItem={{
-                    "buttonProps": consentModalButtonProps,
-                    "text": t("cookies management")
-                }}
-            />
-        );
-    }
-
-    return { FooterConsentManagementItem };
-}
-
-function createFooterPersonalDataPolicyItem(params: {
-    personalDataPolicyLinkProps: RegisteredLinkProps | undefined;
-}) {
-    const { personalDataPolicyLinkProps } = params;
-
-    function FooterPersonalDataPolicyItem() {
-        const { t } = useTranslation();
-
-        if (personalDataPolicyLinkProps === undefined) {
-            throw new Error(
-                [
-                    "You should provide a personalDataPolicyLinkProps to createGdprApi if",
-                    "you want to add a link to the personal data policy in the footer"
-                ].join(" ")
-            );
-        }
-
-        return (
-            <FooterBottomItem
-                bottomItem={{
-                    "text": t("personal data"),
-                    "linkProps": personalDataPolicyLinkProps
-                }}
-            />
-        );
-    }
-    return { FooterPersonalDataPolicyItem };
-}
-
-const { useTranslation, addGdprTranslations } = createComponentI18nApi({
-    "componentName": "Gdpr",
-    "frMessages": {
-        /** cspell: disable */
-        "all services pref": "Préférences pour tous les services.",
-        "personal data cookies": "Données personnelles et cookies",
-        "accept all": "Tout accepter",
-        "accept all - title": "Autoriser tous les cookies",
-        "refuse all": "Tout refuser",
-        "refuse all - title": "Refuser tous les cookies",
-        "accept": "Accepter",
-        "refuse": "Refuser",
-        "confirm choices": "Confirmer mes choix",
-        "about cookies": (p: { hostname: string }) => `À propos des cookies sur ${p.hostname}`,
-        "welcome message": (p: {
-            personalDataPolicyLinkProps: RegisteredLinkProps | undefined;
-        }) => {
-            const { Link } = getLink();
-            return (
-                <>
-                    Bienvenue ! Nous utilisons des cookies pour améliorer votre expérience et les
-                    services disponibles sur ce site.
-                    {p.personalDataPolicyLinkProps !== undefined && (
-                        <>
-                            {" "}
-                            Pour en savoir plus, visitez la page{" "}
-                            <Link {...p.personalDataPolicyLinkProps}>
-                                Données personnelles et cookies
-                            </Link>
-                            .
-                        </>
-                    )}
-                    &nbsp; Vous pouvez, à tout moment, avoir le contrôle sur les cookies que vous
-                    souhaitez activer.
-                </>
-            );
-        },
-        "customize": "Personnaliser",
-        "customize cookies - title": "Personnaliser les cookies",
-        "consent modal title": "Panneau de gestion des cookies",
-        "cookies management": "Gestion des cookies",
-        "personal data": "Données personnelles",
-        "preferences for all services": (p: {
-            personalDataPolicyLinkProps: RegisteredLinkProps | undefined;
-        }) => {
-            const { Link } = getLink();
-            return (
-                <>
-                    Préférences pour tous les services.
-                    <br />
-                    {p.personalDataPolicyLinkProps !== undefined && (
-                        <Link {...p.personalDataPolicyLinkProps}>
-                            Données personnelles et cookies
-                        </Link>
-                    )}
-                </>
-            );
-        },
-        "see more details": "Voir plus de détails",
-        "mandatory cookies": "Cookies obligatoires",
-        "mandatory cookies - description":
-            "Ce site utilise des cookies nécessaires à son bon fonctionnement qui ne peuvent pas être désactivés.",
-        "confirm my choices": "Confirmer mes choix"
-        /** cspell: enable */
-    }
-});
-
-addGdprTranslations({
-    "lang": "en",
-    "messages": {
-        "all services pref": "Preferences for all services.",
-        "personal data cookies": "Personal data and cookies",
-        "accept all": "Accept all",
-        "accept all - title": "Accept all cookies",
-        "refuse all": "Refuse all",
-        "refuse all - title": "Refuse all cookies",
-        "accept": "Accept",
-        "refuse": "Refuse",
-        "confirm choices": "Confirm my choices",
-        "about cookies": p => `About cookies on ${p.hostname}`,
-        "welcome message": p => {
-            const { Link } = getLink();
-            return (
-                <>
-                    Welcome to our website! We use cookies to improve your experience and the
-                    services available services available on this site.
-                    {p.personalDataPolicyLinkProps !== undefined && (
-                        <>
-                            To learn more, visit the{" "}
-                            <Link {...p.personalDataPolicyLinkProps}>
-                                "Personal Data and Cookies"
-                            </Link>{" "}
-                            page.
-                        </>
-                    )}
-                    &nbsp; You can, at any time, have control over which cookies you wish to enable
-                    at any time.
-                </>
-            );
-        },
-        "customize": "Customize",
-        "customize cookies - title": "Customize cookies",
-        "consent modal title": "Cookie management panel",
-        "cookies management": "Cookies management",
-        "preferences for all services": (p: {
-            personalDataPolicyLinkProps: RegisteredLinkProps | undefined;
-        }) => {
-            const { Link } = getLink();
-            return (
-                <>
-                    Preferences for all services <br />
-                    {p.personalDataPolicyLinkProps !== undefined && (
-                        <Link {...p.personalDataPolicyLinkProps}>Personal data and cookies</Link>
-                    )}
-                </>
-            );
-        },
-        "see more details": "See more details",
-        "mandatory cookies": "Mandatory cookies",
-        "mandatory cookies - description":
-            "This site uses cookies necessary for its proper functioning which cannot be deactivated.",
-        "confirm my choices": "Confirm my choices"
-    }
-});
-
-export { addGdprTranslations };
