+++ src/gdpr/ConsentBannerAndConsentManagement/index.ts
@@ -1,0 +1,2 @@
+export * from "./ConsentBannerAndConsentManagement";
+export { addGdprTranslations } from "./translation";
